Version 4.3.61.10 (cherry-pick)
Merged 87a27e42eaf7fc74c2a7b81e4d8261265b9ce326

MIPS64: Unbreak cross build.

R=machenbach@chromium.org, paul.lind@imgtec.com, jkummerow@chromium.org
BUG=

Review URL: https://codereview.chromium.org/1098623004

Cr-Commit-Position: refs/branch-heads/4.3@{#13}
Cr-Branched-From: f5c0a23a505616796a628d64f4ffe377d1fc4bcf-refs/heads/4.3.61@{#1}
Cr-Branched-From: 0a7d4f496a554028de0ab5a963c3a004e693b4cb-refs/heads/master@{#27508}

+++ src/mips64/macro-assembler-mips64.cc
@@ -3240,7 +3240,8 @@ void MacroAssembler::PushStackHandler() {
 void MacroAssembler::PopStackHandler() {
   STATIC_ASSERT(StackHandlerConstants::kNextOffset == 0);
   pop(a1);
-  Daddu(sp, sp, Operand(StackHandlerConstants::kSize - kPointerSize));
+  Daddu(sp, sp, Operand(static_cast<int64_t>(StackHandlerConstants::kSize -
+                                             kPointerSize)));
   li(at, Operand(ExternalReference(Isolate::kHandlerAddress, isolate())));
   sd(a1, MemOperand(at));
 }
